requirements.txt
alembic==1.0.7
aniso8601==4.1.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
